refactor(#518): extract publish button component
